requirements.txt
flask
pillow
json
gunicorn